[RENAME] image structure variable

+++ src/modules/saturation.h
@@ -1,6 +1,6 @@
 #ifndef SATURATION_H
 #define SATURATION_H
 
-void saturation(Image *img, double change);
+void saturation(Image *img, float change);
 
 #endif
